Fixed major bug with HistoricalTicker queries, from / to -.

+++ app/src/main/java/com/example/cryptomarket/data/coinsapi/ticker/HistoricalTicker.kt
@@ -9,7 +9,7 @@ data class HistoricalTicker(
     @Json(name = "price")
     val price: Double,
     @Json(name = "volume_24h")
-    val volume24h: Int,
+    val volume24h: Long,
     @Json(name = "market_cap")
-    val marketCap: Int
+    val marketCap: Long
 )
